+++ R/presets.R
@@ -23,7 +23,7 @@
 
 r_vanilla <- function(func, args = list(), libpath = character(),
                       repos = c(CRAN = "@CRAN@"), cmdargs = "--slave",
-                      system_profile = FALSE, user_profile = FALSE,
+                      system_profile = FALSE, user_profile = "project",
                       env = character(), ...) {
 
   r(func, args = args, libpath = libpath, repos = repos,